Screenshot extension: make sure we use a compatible cairo surface - SL #4104
To create the screenshot surface, use
gdk_window_create_similar_surface instead.  In toolkit
(Activity.get_preview) we do the same.  It creates a more compatible
surface.  Otherwise we are hardcoding the format, which showed to fail
in XO-4 hardware.

To set the window surface as the source surface on the new context,
use gdk_cairo_set_source_window instead.  This way there is no need to
create a new cairo context to get the surface, use the gdk window
directly instead.

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>
Signed-off-by: Ajay Garg <ajay@activitycentral.com>