Tabbinghandler: port the tabbinghandler to the new GTK+3 API, SL #3895
gdk-keyboard-grab [1] and gdk-pointer-grab [2] have been deprecated,
we have to use gdk_device_grab [3] now in order to grab the events. This
patch does the port to the new API.

We also switch to gdk_window_get_device_position instead of
gdk_window_get_pointer [4].

This patch needs as well a fix for the keygrabber (toolkit) to handle
XI2 key events in order to work.

[1] http://developer.gnome.org/gdk3/stable/gdk3-General.html#gdk-keyboard-grab
[2] http://developer.gnome.org/gdk3/stable/gdk3-General.html#gdk-pointer-grab
[3] http://developer.gnome.org/gdk3/stable/GdkDevice.html#gdk-device-grab
[4] http://developer.gnome.org/gdk3/stable/gdk3-Windows.html#gdk-window-get-pointer

Signed-off-by: Simon Schampijer <simon@laptop.org>
Reviewed-by: Gonzalo Odiard <gonzalo@laptop.org>
Tested-by: Gonzalo Odiard <gonzalo@laptop.org>
Signed-off-by: Ajay Garg <ajay@activitycentral.com>