Add support for selecting new XO icon skin
This patch adds the ability fo the Sugar user to select a new skin for
the XO icon from a collection of icons in ~/.sugar/skins. It is an
extension of the aboutme panel, where the user is also able to change
colors and the nick.

This patch contains the modifications to model.py and view.py. It does
not include the sample icons.

Caveat: We need to discuss the mechanism for actually changing the
icon, which is currently stored in /usr/share/icons/sugar/, which is
read-only to Sugar users. Two ideas have been discussed so far: (1) to
make the xo-computer.svg icon read-write; and (2) to add
~/.sugar/icons/ to the head of the icon-themes list, so if a modified
xo-computer is found there, it will be used to override the on ein
/usr/share. This code assumes solution #1 above.

For that reason try (for now) to do:
chmod -R 777 /usr/share/icons/sugar/scalable/device/computer-xo.svg