sdxo#1143: "Hidden-SSIDs" feature.
Note that following are the changes from the earlier implementation:

a)
[UI] No explicit "Save" button.
Instead, the "tick" toolbar button will do this job (taking into proper considerations for the "Cancel Changes",
"Later" and "Restart" scenarios). This keeps it consistent with the rest of the sections in the "Network" panel.

b)
[UI] No explicit "Please reboot your XO to effect changes" message.
Instead the top-toolbar gives this message when the "tick" button is clicked.
This keeps it consistent with the rest of the sections in the "Network" panel.

c)
[BACKEND] The "iwconfig" network-scanning will now be done at "sugar" (re)start, instead of "system"" (re)start.

Thankfully, the c)  change now also means that the "restart"
notification in "My Settings" -> "Network", is now also consistent, as
the restart here refers to "sugar" restart. Thus, "sugar" restart is
what is now all that is needed to make the hidden wireless networks appear in the
Neighborhood-View.

The only caveat: "sudo" must be available for the "olpc" user (without needing a password).