Journal sorting button: hide/show on click or touch
Using the 'toggle_palette' property of the Palette invoker
we get the behaviour for hiding/showing the Palette when
clicking or touching the button. We need to set the
'hide_tooltip_on_click' property of the ToolButton to False
in that case. This is the same behaviour as the Frame icons,
see the WirelessDeviceView for a similar case of a ToolButton
with that behaviour.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>
Signed-off-by: Ajay Garg <ajay@activitycentral.com>